refactor(api): flatten API package layout
