+++ PROJECT
@@ -1,7 +1,6 @@
 domain: cybozu.com
 layout:
 - go.kubebuilder.io/v4
-multigroup: true
 projectName: accurate
 repo: github.com/cybozu-go/accurate
 resources:
@@ -12,7 +11,7 @@ resources:
   domain: cybozu.com
   group: accurate
   kind: SubNamespace
-  path: github.com/cybozu-go/accurate/api/accurate/v2
+  path: github.com/cybozu-go/accurate/api/v2
   version: v2
   webhooks:
     defaulting: true
