+++ cmd/kubectl-accurate/sub/sub_create.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	accuratev2 "github.com/cybozu-go/accurate/api/accurate/v2"
+	accuratev2 "github.com/cybozu-go/accurate/api/v2"
 	"github.com/spf13/cobra"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
