+++ internal/controller/namespace_controller.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"path"
 
-	accuratev2 "github.com/cybozu-go/accurate/api/accurate/v2"
+	accuratev2 "github.com/cybozu-go/accurate/api/v2"
 	"github.com/cybozu-go/accurate/internal/indexing"
 	utilerrors "github.com/cybozu-go/accurate/internal/util/errors"
 	corev1 "k8s.io/api/core/v1"
