+++ internal/controller/suite_test.go
@@ -1,3 +1,5 @@
+//go:build envtest
+
 package controllers
 
 import (
@@ -19,7 +21,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	accuratev2 "github.com/cybozu-go/accurate/api/accurate/v2"
+	accuratev2 "github.com/cybozu-go/accurate/api/v2"
 	//+kubebuilder:scaffold:imports
 )
