+++ internal/indexing/constants.go
@@ -6,5 +6,5 @@ const (
 	NamespaceTemplateKey = "namespace.template"
 	PropagateKey         = "resource.propagate"
 	SubNamespaceNameKey  = "subnamespace.name"
-	PropagateAny         = "any" // defined as an in-memory index value
+	propagateAny         = "any" // defined as an in-memory index value
 )
