+++ internal/webhook/v2/allow-cascade-delete/cascade_delete_test.go
@@ -1,3 +1,5 @@
+//go:build envtest
+
 package hooks_allow_cascade_delete
 
 import (
@@ -10,7 +12,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	accuratev2 "github.com/cybozu-go/accurate/api/accurate/v2"
+	accuratev2 "github.com/cybozu-go/accurate/api/v2"
 )
 
 var _ = Describe("Webhook allow cascade delete", func() {
