+++ internal/webhook/v2/namespace_test.go
@@ -1,9 +1,11 @@
+//go:build envtest
+
 package v2
 
 import (
 	"context"
 
-	accuratev2 "github.com/cybozu-go/accurate/api/accurate/v2"
+	accuratev2 "github.com/cybozu-go/accurate/api/v2"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
